+++ dags/tests/test_datasources_to_xnat/TEST_settings.yml
